requirements.txt
torch
torch-geometric
pandas
numpy
networkx
pyyaml
scikit-learn
matplotlib
jupyter
